pyproject.toml
[tool.poetry]
name = "insider-detection-blockchain"
version = "0.1.0"
description = "Blockchain insider activity detection system using RAG and multi-agent architecture"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# LLM & Agent Framework
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langgraph = "^0.2.0"
langsmith = "^0.1.0"

# Embedding Model
sentence-transformers = "^3.0.0"
transformers = "^4.45.0"
torch = "^2.4.0"

# Vector Database
qdrant-client = "^1.11.0"

# Graph Database
neo4j = "^5.24.0"

# OLAP Database
clickhouse-connect = "^0.7.0"

# Caching & Task Queue
redis = "^5.0.0"
celery = "^5.4.0"

# API Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"

# Ethereum
web3 = "^7.2.0"

# Observability
prometheus-client = "^0.21.0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
structlog = "^24.4.0"

# Utilities
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
tenacity = "^9.0.0"
orjson = "^3.10.0"

[tool.poetry.group.airflow.dependencies]
apache-airflow = "^2.10.0"
apache-airflow-providers-celery = "^3.8.0"
apache-airflow-providers-redis = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
pre-commit = "^3.8.0"
ipython = "^8.27.0"
jupyter = "^1.1.0"

[tool.poetry.group.test.dependencies]
faker = "^28.0.0"
hypothesis = "^6.112.0"
testcontainers = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "neo4j.*",
    "clickhouse_connect.*",
    "sentence_transformers.*",
    "celery.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]
